Put class declaration into header file.
Also change the DRIVER_MODIFICATION number even though I'm not sure if
this should be done since this shouldn't change anything in the driver
per se.
